requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
blinker==1.6.2
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ConfigArgParse==1.7
Django==4.2.4
djangorestframework==3.14.0
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.7.0
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.16.1
MarkupSafe==2.1.3
msgpack==1.0.5
mysqlclient==2.2.0
# numpy==1.24.4
# pandas==2.0.3
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
validators==0.21.2
Werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
